refactor: centralize date formatting utils
