+++ src/utils/dateUtils.js
@@ -1,0 +1,23 @@
+export function formatDate(dateString) {
+  if (!dateString) return "";
+  const dateObj = new Date(dateString);
+  if (isNaN(dateObj)) return dateString;
+  const dd = String(dateObj.getDate()).padStart(2, "0");
+  const mm = String(dateObj.getMonth() + 1).padStart(2, "0");
+  const yyyy = dateObj.getFullYear();
+  const hh = String(dateObj.getHours()).padStart(2, "0");
+  const min = String(dateObj.getMinutes()).padStart(2, "0");
+  return `${dd}-${mm}-${yyyy} ${hh}:${min}`;
+}
+
+export function formatLongDate(dateStr) {
+  if (!dateStr) return "";
+  const d = new Date(dateStr);
+  if (isNaN(d)) return dateStr;
+  const weekday = d.toLocaleDateString("es-ES", { weekday: "long" });
+  const day = d.toLocaleDateString("es-ES", { day: "2-digit" });
+  const month = d.toLocaleDateString("es-ES", { month: "long" });
+  const year = d.toLocaleDateString("es-ES", { year: "numeric" });
+  const capitalizedWeekday = weekday.charAt(0).toUpperCase() + weekday.slice(1);
+  return `${capitalizedWeekday} ${day} de ${month} de ${year}`;
+}
